### 2017-08-07 #140 @beanumber
etl: framework for medium data
Facilitates predictable and pipeable ETL (extract-transform-load) operations for publicly-accessible medium data sets

```
Package: etl
Title: Extract-Transform-Load Framework for Medium Data
Version: 0.3.6
    person("Ben", "Baumer", email = "ben.baumer@gmail.com",
      role = c("aut", "cre")),
    person("Carson", "Sievert", email = "cpsievert1@gmail.com", role = "ctb"))
Maintainer: Ben Baumer <ben.baumer@gmail.com>
Description: A predictable and pipeable framework for performing ETL
    (extract-transform-load) operations on publicly-accessible medium-sized data
    set. This package sets up the method structure and implements generic
    functions. Packages that depend on this package download specific data sets
    from the Internet, clean them up, and import them into a local or remote
    relational database management system.
License: CC0
    DBI,
    datasets,
    downloader,
    utils
    R (>= 2.10),
    dplyr
    airlines,
    dbplyr,
    RSQLite,
    RPostgreSQL,
    RMySQL,
    MonetDBLite,

```

http://github.com/beanumber/etl

reproducibility, because the extensions of this package will lead to reproducible medium data set used in research
data retrieval, since the extensions of this package download data
data munging, since the extensions of this package transform raw data into CSVs

R developers for the `etl` package itself
R users for `etl`-dependent packages

No. This package depends heavily on `dplyr` and `dbplyr`, but it provides functionality specific to the ETL process that is not present in either.
