### 2017-08-08 #140 @maelle
Thanks a lot for your submission @beanumber! We (rOpenSci onboarding editors) discussed the fit of the package and don't think it's in scope. In particular we couldn't see the scientific application of the package, why it'd lead to more reproducibility than other approaches. If you disagree with this decision feel free to provide us with a more specific/descriptive explanation.

Don't hesitate to submit other packages in the future, potentially starting with a pre-submission enquiry in this same repo.
