### 2017-08-08 #140 @maelle
Thanks both but it's still unclear how this improves _reproducibility_ compared to existing approaches such as e.g. [this one for big datasets](https://journal.r-project.org/archive/2017/RJ-2017-026/index.html)? (data provenance, versioning, etc.)

Our saying the package is out-of-scope doesn't mean it's useless, of course!
