### 2017-08-24 #140 @beanumber
I don't know if this will change your position @maelle, but I've posted the [long-form article for this](https://arxiv.org/abs/1708.07073) on the arXiv. The manuscript explains the package itself and the purpose of the package in far greater detail.
