### 2017-08-24 #140 @maelle
Thanks @beanumber, the editors position was because we couldn't see how the package helps *reproducibility* compared to existing approaches. Is there a part of the article dealing more specifically with this? In any case good work.
