### 2017-08-24 #140 @beanumber
Section 2 addresses this. You might be interested in Section 4.2 to see an example of how this might work in practice.
