### 2017-08-27 #140 @maelle
@beanumber, thanks for providing us the link to your manuscript.

After discussion within the editorial team we still think that albeit very useful `etl` is out of scope for rOpenSci and here is why:

* It is a general data manipulation tool, not specifically aimed at retrieving or extracting a data type or source

* The reproducibility packages that are in scope as per our policies are "Tools that facilitate reproducible research. This includes packages that facilitate use of version control, provenance tracking, automated testing of data inputs and statistical outputs, citation of software and scientific literature.",

which isn't the case for `etl`. `etl` doesn't help tracking provenance or versions of a dataset.

We however encourage the development of your package and your communication efforts for making it better known. We suggest you submit it to the R Journal.

Thanks again, and don't hesitate to ask any question.
